BulkUpdatable: refactor test to loop through types
I'm planning on adding a bunch more types to the tests, so wanted to
make it easier to add them without needing to come up with witty names
for each.

A side effect of this change is that I discovered that `hstore` wasn't
being tested because it was getting overridden by the `jsonb` column
with the same name. I added a line to `spec_helper.rb` to create the
`hstore` extension if it doesn't already exist, being that it didn't
exist for me in the test database at first.