BulkUpdatable: extract specs to shared examples
I'm going to be adding a lot more types to be tested against here, so
wanted to make them a little more generic. This adds a new set of shared
examples and runs them against each type. I added a `GENERATORS` hash to
generate data for each type and a couple of helper methods in the test.
One weird thing is that in order to access `type` in the methods, I
needed to set it in a `let`, otherwise it would throw a `NoMethodError`.