BulkUpdatable: remove manual type cast map (#15)
So it turns out we can apparently use `column.sql_type` for the cast.
How cool is that? I also added a few more types to the tests, and a list
of other types we might want to test at some point. Not sure how
important it is, though, to test all of them, being that it seems like
the generalized `sql_type` might just cover our butts on all types.