BulkUpdatable: add support for jsonb_accessor (#18)
So the big thing with `jsonb_accessor` is that it adds the virtual
attributes to `changes` in addition to the column:

```rb
class Step < ApplicationRecord
  jsonb_accessor(:settings, skippable: :boolean)
end

step = Step.first
step.changes # => {}
step.skippable = true
step.changes # => {"settings"=>[{"skippable"=>nil}, {"skippable"=>true}], "skippable"=>[nil, true]}
```

Thanks to the fact that It adds *both*, we can simply ignore the virtual
attribute and take the column name. Rails has a convenient
`column_names` method which generally returns the names of the columns
from the database. Except that the older `jsonb_accessor ~> 0.3.3` causes
a fake column to be added to the list. I was able to determine that
these columns don't have a `sql_type` like the rest of the columns, so
we just filter those out.

Note: I needed to add a handful of conditional checks to support the
`jsonb_accessor ~> 0.3.3` for Rails 4.2, and `jsonb_accessor >= 1.0.0`
for Rails >= 5. Rather than checking the `jsonb_accessor` version
directly, I opted to check against `ActiveRecord::VERSION::MAJOR`. I
figured this would be a little easier to search for when we decide to
drop support for Rails 4.2.