BulkUpdatable: clear changes after query (#20)
Pretty simple. When we call `#save` on a record, ActiveRecord clears
changes, so it makes sense to do the same when bulk updating. One thing
of note is that `clear_changes_information` clobbers all change
information. `#save` moves `changes` to `previous_changes`, which ought
to be possible via the `#changes_applied` method, but there appears to
be some weird caching going on preventing this from working, so we went
with `#clear_changes_information`, which clears both `changes` and
`previous_changes`. This may be something we can revisit in Rails 5.2 or
later, as it appears this logic has shifted quite a bit.

Rails 5.1:
https://github.com/rails/rails/blob/v5.1.6/activerecord/lib/active_record/attribute_methods/dirty.rb#L71-L83

Rails 5.2, no change tracking methods:
https://github.com/rails/rails/blob/v5.2.0/activerecord/lib/active_record/attribute_methods/dirty.rb