Combine changes across instances of same object
**What**

Combines the attribute changes across multiple instances of the same
object.

The attribute changes will be on each of the objects' instances, but the
update query will only run on one of those instances to insure the data is
not lost.

**Why**

It's easily possible for a user to include multiple instances of the same
object in a bulk_update transaction. Unfortunately, this is not currently
handled very well. Essentially, what currently happens is the changes that
end up getting persisted in the order of `object.column_names`, so there
is lost data.

[ch1293]