Upgrades: fix deprecations in Ruby 2.7
Ruby complains about implicitly passing a hash as keyword arguments. If
we want to pass in a hash that will be slurped up on the other side with
`**`, then we need to splat it out with `**` when we're calling it. This
fixes deprecation warnings to this effect, though the fixes should be
backwards compatible.

**Notes**

* I removed the `bundler` version from the `gemspec` file. I don't think
  it makes any sense to have in there.
* I added more ruby versions to the `.travis.yml` section. We can
  probably deprecate some of the older versions we're testing against.