Upgrades: fix keyword arguments across versions (#29)
Fixes #28. Ruby 2.6 splat arguments behave differently from Ruby 2.7
splat arguments which behave (slightly) differently from Ruby 3.0 splat
arguments. It turns out it's hard to handle all of them gracefully, but
thankfully a little gem called [`ruby2_keywords`][rbkw] is provided by
Ruby core for us to transition smoothly. For Ruby versions 2.6 and
earlier, it is a no-op. For Ruby versions 2.7 and up, the method is
already defined in core Ruby, in which cases the gem does nothing. The
core version internally translates the arguments to Ruby 3 style keyword
arguments. You can read all about the complexities [here][blog]. There
are several other suggested ways of handling the issue, but this seemed
to be the cleanest. Once Ruby <= 2.6 support is dropped, this should be
able to be removed.

[rbkw]: https://github.com/ruby/ruby2_keywords
[blog]: https://eregon.me/blog/2019/11/10/the-delegation-challenge-of-ruby27.html