Notifier: add ability to configure with arbitrary callable strategy (#31)
**What**
Updates JunkDrawer::Notifier to accept any arbitrary callable object
(i.e. anything that `responds_to?` call) as a strategy.

**Why**
Currently, there are only a few pre-defined strategies that the Notifier
can be configured with. This addition provides a lot more flexibility,
allowing users of the gem to define their own custom strategies on the
fly.